refactor(agentic-sorting): remove Replicate, rename Lab.tsx, re-tag Track
Legibility pass (1/4): the Replicate panel duplicated the Lab's Monte-Carlo, so
remove it (panel + 5 state vars + runReplicate) now that the Lab covers
'same params, different instances'. Rename Lab.tsx -> LabResults.tsx to end the
'two Labs' name clash. Re-tag the Track agent panel arch 'lab' -> 'readout' (it
is a readout, not a batch experiment), de-crowding the rail's Analyze tier.
