+++ src/animations/AgenticSorting/LabResults.tsx
